refactor: extract common dialog component
