+++ .changeset/quiet-donkeys-punch.md
@@ -1,0 +1,5 @@
+---
+'@backstage/plugin-kubernetes-react': patch
+---
+
+Extracted common dialog component.
